requirements.txt
torch>=2.0.0
transformers>=4.17.0
tqdm>=4.64.0
wandb>=0.12.11
numpy>=1.22.3
matplotlib>=3.5.0
scikit-learn>=1.0.0
pyyaml>=6.0
dataclasses>=0.5